Function npath::normalize

pub fn normalize<P: AsRef<Path>>(path: P) -> PathBuf

Returns the shortest equivalent path by pure lexical processing.

It applies the following rules:

replaces multiple / with a single one,

eliminates each . path name element (the current directory),

eliminates each inner .. path name element (the parent directory) along with the non-.. element that precedes it,

eliminates .. elements that begin a rooted path, that is, replace /.. by / at the beginning of a path

Differences with fs::canonicalize

This function does not touch the filesystem, ever:

it does not resolve symlinks,

it does not check if files/directories exists,

if the given path is relative, it returns a relative path, etc.

If /a/b is a symlink to /d/e, then /a/b/../c is /d/c (and not /a/c that this function will return).

Example

use std::path::Path;
use npath::normalize;

fn main() {
    assert_eq!(normalize("usr/lib"), Path::new("usr/lib"));
    assert_eq!(normalize("usr//lib"), Path::new("usr/lib"));
    assert_eq!(normalize("usr/lib/."), Path::new("usr/lib"));
    assert_eq!(normalize("usr/lib/gcc/.."), Path::new("usr/lib"));
    assert_eq!(normalize("/../usr/lib"), Path::new("/usr/lib"));
    assert_eq!(normalize("/../usr/bin/../././/lib"), Path::new("/usr/lib"));
}