Trait npath::NormPathExt

pub trait NormPathExt {
    fn absolute(&self) -> Result<PathBuf>;

    fn base(&self) -> &Path;

    fn dir(&self) -> &Path;

    fn is_inside<P: AsRef<Path>>(&self, base: P) -> bool;

    fn normalized(&self) -> PathBuf;

    fn lexical_join<P: AsRef<Path>>(&self, path: P) -> PathBuf;

    fn relative_to<P: AsRef<Path>>(&self, base: P) -> Option<PathBuf>;

    fn resolved(&self) -> Result<PathBuf>;

    fn rooted_join<P: AsRef<Path>>(&self, path: P) -> PathBuf;

    fn try_rooted_join<P: AsRef<Path>>(&self, path: P) -> Result<PathBuf>;
}

Extension trait for Path.

Required methods

fn absolute(&self) -> Result<PathBuf>

Returns the absolute equivalent of self.

If self is absolute, it it is returned as is.

If self is relative, it is lexically joined to std::env::current_dir.

Examples

use std::path::{Path, PathBuf};
use npath::NormPathExt;

assert_eq!(Path::new("/usr").absolute().unwrap(), PathBuf::from("/usr"));

if let Ok(cwd) = std::env::current_dir() {
    assert_eq!(Path::new("lib").absolute().unwrap(), cwd.lexical_join("lib"));
}

fn base(&self) -> &Path

Returns the last path component of self.

See basename(3).

Differences with Path::file_name

Always returns a path (eventually /, ., or .., when Path::file_name returns None).

Returns a Path instead of an OsStr.

Examples

use std::path::Path;
use npath::NormPathExt;

assert_eq!(Path::new("/usr/lib").base(), Path::new("lib"));
assert_eq!(Path::new("/usr/").base(),    Path::new("usr"));
assert_eq!(Path::new("usr/.").base(),    Path::new("."));
assert_eq!(Path::new("usr").base(),      Path::new("usr"));
assert_eq!(Path::new("/").base(),        Path::new("/"));
assert_eq!(Path::new(".").base(),        Path::new("."));
assert_eq!(Path::new("..").base(),       Path::new(".."));

fn dir(&self) -> &Path

Returns self up to, but not including, the final component.

See dirname(3).

Differences with Path::parent

Always returns a path (/ when Path::parent returns None).

Examples

use std::path::Path;
use npath::NormPathExt;

assert_eq!(Path::new("/usr/lib").dir(), Path::new("/usr"));
assert_eq!(Path::new("/usr/").dir(),    Path::new("/"));
assert_eq!(Path::new("usr/.").dir(),    Path::new("usr"));
assert_eq!(Path::new("usr").dir(),      Path::new("."));
assert_eq!(Path::new("/").dir(),        Path::new("/"));
assert_eq!(Path::new(".").dir(),        Path::new("."));
assert_eq!(Path::new("..").dir(),       Path::new("."));

fn is_inside<P: AsRef<Path>>(&self, base: P) -> bool

Returns whether self is lexically inside base.

self is considered "lexically inside" base (or a descendant of base) if and only if:

self and base are both relative, or both absolute.

self does not have any component outside of base.

Both paths are normalized before being compared.

Limitations

When processing relative paths, if it is necessary to know the absolute path of either base or self, this method returns false:

If base is absolute and self is relative, or the opposite, there is no way to know where base points to relative to self without using the absolute path of the CWD.

If base and self are relative, and base is re-entrant (a parent component .. is followed by a normal component), it is not possible to know whether base re-enters the directory it just left, or if it branches off to a sibling directory, without using the absolute path of the CWD. (Note that self can be re-entrant as shown in the examples.)

To circumvent these limitations, you can ensure both path are absolute using one of these methods:

NormPathExt::absolute: with the limitations of NormPathExt::normalized.

Path::canonicalize: without the limitations of NormPathExt::normalized, but both paths must exist on the filesystem.

Windows

Paths components are compared in a case-insensitive way only for the ASCII character set (first 128 Unicode characters). For paths containing characters outside of this range, this method could return false even though self is inside base on the filesystem.

Examples

use std::path::Path;
use npath::NormPathExt;

// Both absolute
assert!(Path::new("/srv").is_inside("/"));
assert!(Path::new("/srv").is_inside("/srv"));
assert!(Path::new("/srv/file.txt").is_inside("/srv"));

// Both relative
assert!(Path::new("srv").is_inside("srv"));
assert!(Path::new("srv").is_inside("."));
assert!(Path::new("srv").is_inside(".."));
assert!(Path::new("../srv").is_inside("..")); // self is re-entrent

// One absolute, the other relative
assert!(!Path::new("file.txt").is_inside("/srv"));
assert!(!Path::new("/srv/file.txt").is_inside("srv"));

// self exits base
assert!(!Path::new("/srv/..").is_inside("/srv"));

// base is re-entrent
assert!(!Path::new("srv").is_inside("../foo"));

fn normalized(&self) -> PathBuf

Returns the normalized equivalent of self.

The returned path is the shortest equivalent, normalized by pure lexical processing with the following rules:

Replace repeated / with a single one.

Eliminate . path components (the current directory).

Simplify inner .. path components (the parent directory), including components preceded by a rooted path (replace /.. by /).

Differences with Path::canonicalize

This function does not touch the filesystem, ever:

It does not resolve symlinks.

It does not check if files/directories exist.

If the given path is relative, it returns a relative path.

Limitations

If /a/b is a symlink to /d/e, then for /a/b/../c:

Path::canonicalize returns /d/c.

NormPathExt::normalized returns /a/c.

Examples

use std::path::{Path, PathBuf};
use npath::NormPathExt;

assert_eq!(Path::new("usr/lib").normalized(),                 PathBuf::from("usr/lib"));
assert_eq!(Path::new("usr//lib").normalized(),                PathBuf::from("usr/lib"));
assert_eq!(Path::new("usr/lib/.").normalized(),               PathBuf::from("usr/lib"));
assert_eq!(Path::new("usr/lib/gcc/..").normalized(),          PathBuf::from("usr/lib"));
assert_eq!(Path::new("/../usr/lib").normalized(),             PathBuf::from("/usr/lib"));
assert_eq!(Path::new("/../usr/bin/../././/lib").normalized(), PathBuf::from("/usr/lib"));

fn lexical_join<P: AsRef<Path>>(&self, path: P) -> PathBuf

Returns path lexically joined to self.

Differences with Path::join

If path is absolute, it does not replace self.

Limitations

On Windows, this method can turn paths without UNC prefixes into paths that have one.

use std::path::{Component, Path, PathBuf};
use npath::NormPathExt;

fn has_prefix(path: &Path) -> bool {
    path.components().next().map(|c| matches!(c, Component::Prefix(_))).unwrap_or(false)
}

assert_eq!(
    Path::new(r"\\server").lexical_join(r"c$\file.txt"),
    PathBuf::from(r"\\server\c$\file.txt"),
);

Examples

use std::path::{Path, PathBuf};
use npath::NormPathExt;

assert_eq!(
    Path::new("/usr")
        .lexical_join("bin")   // relative
        .lexical_join("/cat"), // absolute
    PathBuf::from("/usr/bin/cat")
);

assert_eq!(Path::new("/usr").lexical_join(".."), PathBuf::from("/usr/.."));

fn relative_to<P: AsRef<Path>>(&self, base: P) -> Option<PathBuf>

Returns the normalized relative path from base to self.

If self cannot be made relative to base, the method returns None. The returned path, if any, satisfies: base.join(path) == self. Both paths are normalized before being compared, and the returned path is also normalized.

Limitations

The method returns None when it cannot determine the intermediate path components to reach self from base, or fetching the CWD is required:

One path is absolute, but the other is relative.

The paths are relative, base is above self, and they are both above the CWD (knowing its intermediate components is necessary).

To circumvent these limitations, you can ensure both path are absolute using one of these methods:

NormPathExt::absolute: with the limitations of NormPathExt::normalized.

Path::canonicalize: without the limitations of NormPathExt::normalized, but both paths must exist on the filesystem.

Windows

There is an inescapable limitation when the paths refer to different locations (different drives, or network shares): there is no relative path between them because they are not in the same namespace.

Paths components are compared in a case-insensitive way only for the ASCII character set (first 128 Unicode characters). For paths containing characters outside of this range, this method could return None even though there is a relative path from base to self on the filesystem.

Examples

use std::path::{Path, PathBuf};
use npath::NormPathExt;

assert_eq!(Path::new("/usr/lib").relative_to("/usr"), Some(PathBuf::from("lib")));
assert_eq!(Path::new("usr/bin").relative_to("var"),   Some(PathBuf::from("../usr/bin")));

assert_eq!(Path::new("lib").relative_to("/usr"), None);
assert_eq!(Path::new("/usr").relative_to("var"), None);

assert_eq!(Path::new("foo").relative_to("/"),  None);
assert_eq!(Path::new("foo").relative_to(".."), None);

assert_eq!(Path::new("..").relative_to("."), Some(PathBuf::from("..")));
assert_eq!(Path::new(".").relative_to(".."), None);

fn resolved(&self) -> Result<PathBuf>

Returns the normalized equivalent of self, with intermediate symlinks resolved.

The longest prefix of self where each component exist on the filesystem is canonicalized, the remaining path is lexically adjoined, and the result is normalized.

If self is relative and no prefix component exists in the CWD (no prefix can be canonicalized), then the result is relative (it is simply the normalized equivalent of self). Use NormPathExt::absolute to ensure this method returns an absolute path.

Differences with Path::canonicalize

This method works for paths that do not exist on the filesystem.

The path is normalized with the same limitations as NormPathExt::normalized. In particular, a suffix might contain .. components that can replace canonicalized components.

The returned path can be relative.

Differences with NormPathExt::normalized

This method does not perform pure lexical processing, and returns a std::io::Result.

The prefix that is canonicalized has its intermediate symlinks resolved, without the limitations of NormPathExt::normalized.

Limitations

The prefix considered for canonicalization ends at the first component of self that does not exist. The only exception is if the whole path exists, then it is canonicalized in full.

For example, assuming only usr and usr/lib exist, usr/liz/../lib/rust refers to usr/lib/rust, but only usr is canonicalized since neither usr/lib/rust nor usr/liz exist.

NormPathExt::normalized limitations apply in most cases. If you need to get the true canonical path, use Path::canonicalize.

Examples

use std::path::{Path, PathBuf};
use npath::NormPathExt;

assert_eq!(Path::new("usr/lib").resolved().unwrap(),                 PathBuf::from("usr/lib"));
assert_eq!(Path::new("usr//lib").resolved().unwrap(),                PathBuf::from("usr/lib"));
assert_eq!(Path::new("usr/lib/.").resolved().unwrap(),               PathBuf::from("usr/lib"));
assert_eq!(Path::new("usr/lib/gcc/..").resolved().unwrap(),          PathBuf::from("usr/lib"));
assert_eq!(Path::new("/usr/lib").resolved().unwrap(),                PathBuf::from("/usr/lib"));
assert_eq!(Path::new("/../usr/lib").resolved().unwrap(),             PathBuf::from("/usr/lib"));
assert_eq!(Path::new("/usr/bin/../././/lib").resolved().unwrap(),    PathBuf::from("/usr/lib"));

fn rooted_join<P: AsRef<Path>>(&self, path: P) -> PathBuf

Returns path restricted to self.

This methods works as if self was the root directory. The returned path is guaranteed to be lexically inside self.

Differences with NormPathExt::try_rooted_join

It does not return an error since the returned path is always inside self.

Limitations

Same limitations as NormPathExt::normalized.

The result can have a different meaning than what was intended since all the prefix .. components are stripped away.

Examples

use std::path::{Path, PathBuf};
use npath::NormPathExt;

assert_eq!(Path::new("/srv").rooted_join("file.txt"),           PathBuf::from("/srv/file.txt"));
assert_eq!(Path::new("/srv").rooted_join("/file.txt"),          PathBuf::from("/srv/file.txt"));
assert_eq!(Path::new("/srv").rooted_join("../file.txt"),        PathBuf::from("/srv/file.txt"));
assert_eq!(Path::new("/srv").rooted_join("foo/../file.txt"),    PathBuf::from("/srv/file.txt"));
assert_eq!(Path::new("/srv").rooted_join("foo/../../file.txt"), PathBuf::from("/srv/file.txt"));
assert_eq!(
    Path::new("/srv").rooted_join("../srv/file.txt"),
    PathBuf::from("/srv/srv/file.txt")
);

fn try_rooted_join<P: AsRef<Path>>(&self, path: P) -> Result<PathBuf>

Returns path restricted to self.

self and path are converted to absolute paths with NormPathExt::absolute, they are joined with NormPathExt::lexical_join, and the result is normalized with NormPathExt::normalized.

Differences with NormPathExt::rooted_join

If the result points to a location outside of base, this method returns an error.

Examples

use std::path::{Path, PathBuf};
use npath::NormPathExt;

assert_eq!(Path::new("/srv").try_rooted_join("file.txt").unwrap(),        PathBuf::from("/srv/file.txt"));
assert_eq!(Path::new("/srv").try_rooted_join("/file.txt").unwrap(),       PathBuf::from("/srv/file.txt"));
assert_eq!(Path::new("/srv").try_rooted_join("foo/../file.txt").unwrap(), PathBuf::from("/srv/file.txt"));
assert_eq!(
    Path::new("/srv").try_rooted_join("../srv/file.txt").unwrap(),
    PathBuf::from("/srv/file.txt")
);

assert!(Path::new("/srv").try_rooted_join("../file.txt").is_err());
assert!(Path::new("/srv").try_rooted_join("foo/../../file.txt").is_err());

Loading content...